--- a/src/EllipticCurveCryptography/twoPower.lua
+++ b/src/EllipticCurveCryptography/twoPower.lua
@@ -1,0 +1,14 @@
+local twoPower = setmetatable({}, {
+	__index = function(self, index)
+		local value = 2 ^ index
+		self[index] = value
+		return value
+	end,
+})
+
+-- precache
+for index = -512, 512 do
+	local _ = twoPower[index]
+end
+
+return twoPower