--- a/src/EllipticCurveCryptography/util.lua
+++ b/src/EllipticCurveCryptography/util.lua
@@ -4,21 +4,26 @@ util.byteTableMT = {
 	__tostring = function(a)
 		return string.char(table.unpack(a))
 	end,
+
 	__index = {
 		toHex = function(self)
-			return ("%02x"):rep(#self):format(table.unpack(self))
+			return string.format(string.rep("%02x", #self), table.unpack(self))
 		end,
+
 		isEqual = function(self, t)
 			if type(t) ~= "table" then
 				return false
 			end
+
 			if #self ~= #t then
 				return false
 			end
+
 			local ret = 0
-			for i = 1, #self do
-				ret = bit32.bor(ret, bit32.bxor(self[i], t[i]))
+			for index, value in ipairs(self) do
+				ret = bit32.bor(ret, bit32.bxor(value, t[index]))
 			end
+
 			return ret == 0
 		end,
 	},
@@ -29,14 +34,16 @@ function util.stringToByteArray(str)
 		return {}
 	end
 
-	if #str < 7000 then
-		return table.pack(str:byte(1, -1))
+	local length = #str
+	if length < 7000 then
+		return table.pack(string.byte(str, 1, -1))
 	end
 
-	local arr = table.create(#str)
-	for i = 1, #str do
+	local arr = table.create(length)
+	for i = 1, length do
 		arr[i] = string.byte(str, i)
 	end
+
 	return arr
 end