--- a/stylua.toml
+++ b/stylua.toml
@@ -1,0 +1,3 @@
+# column_width = 248347834343
+column_width = 120
+quote_style = "ForceDouble"