I _think_ it still works (running the tests says it passed, so I'm gonna take that as a yes).

Here's the results I got:

```
Ecc:
	Dataset: 500 payloads of 100 bytes of random data.
	Results:
		Encrypt & Sign took 5465.16ms in total with a 10.93ms avg.
		Decrypt & Verify took 8538.27ms in total with a 17.08ms avg.

FastEcc:
	Dataset: 500 payloads of 100 bytes of random data.
	Results:
		Encrypt & Sign took 1912.72ms in total with a 3.83ms avg.
		Decrypt & Verify took 4156.13ms in total with a 8.31ms avg.
```

[Benchmarker.zip](https://github.com/boatbomber/EncryptedNet/files/8128752/Benchmarker.zip)